Make nduplicate time-dependent
`nduplicate` in gate devices has been a constant variable, and this
update makes `nduplicate` time-dependent. This change is needed for a
recent requirement to adjust the number of gates over time.

DSM2-1098

+++ dsm2/src/fixed/buffer_input_grid.f90
@@ -54,7 +54,7 @@ subroutine buffer_input_grid() !process_text_grid_input()
        character(len=32) :: device_name
        character(len=16) :: default_op
        character(len=8) :: struct_name
-       integer :: nduplicate
+       real(8) :: nduplicate
        !real(8) :: width
        !real(8) :: elev
        real(8) :: height
